+++ home/src/app/home/home.component.ts
@@ -1,4 +1,4 @@
-import {Component} from '@angular/core';
+import { Component } from '@angular/core';
 import {UserAccount} from '../user/user-account';
 import {AngularFireAuth} from '@angular/fire/auth';
 import {AngularFirestore} from '@angular/fire/firestore';
@@ -7,7 +7,9 @@ import {of} from 'rxjs';
 import {DrupalService} from '../services/drupal.service';
 import {formatNumber} from '@angular/common';
 import { NgxUiLoaderService } from 'ngx-ui-loader';
-import { CookiesService } from '../services/cookies.service';
+import { AuthService } from '../services/auth/auth.service';
+import { CookieService } from 'ngx-cookie';
+import { isPlatformBrowser } from '@angular/common';
 
 @Component({
   selector: 'app-home',
@@ -82,20 +84,36 @@ export class HomeComponent {
     private angularFireAuth: AngularFireAuth,
     private angularFirestore: AngularFirestore,
     private drupalService: DrupalService,
-    private ngxService: NgxUiLoaderService
+    private ngxService: NgxUiLoaderService,
+    private authService: AuthService,
+    private cookieService: CookieService
   ) {
-    this.ngxService.start();
+
+    this.authService.getDrupalToken()
+    .subscribe(_ => {
+      this.getHomePage();
+    });
+    this.getHomePage();
+    const search = new URLSearchParams(window.location.search);
+    if (search.get('source')) {
+      this.isOnWebView = true;
+    }
+  }
+
+  getHomePage():void {
     this.drupalService.getPage('home')
     .pipe(map(sections => this.fillPlaceholders(sections)))
     .subscribe(sections => {
       this.sections = sections;
-      this.ngxService.stop();
+      //this.ngxService.stop();
     });
-    const path = CookiesService.getValue('LOGIN_REDIRECTION');
-    if(path && path.includes('#') && CookiesService.getValue('DRUPAL_SESSION')) {
-      const section = document.getElementById(path.replace('#', ''));
-      if(section)
-        window.scrollTo(0, section.offsetTop);
+    if (isPlatformBrowser) {
+      const path = this.cookieService.get('LOGIN_REDIRECTION');
+      if(path && path.includes('#') && this.cookieService.get('DRUPAL_SESSION')) {
+        const section = document.getElementById(path.replace('#', ''));
+        if(section)
+          window.scrollTo(0, section.offsetTop);
+      }
     }
     const search = new URLSearchParams(window.location.search);
     if (search.get('source')) {
@@ -107,13 +125,13 @@ export class HomeComponent {
     for (const section of sections) {
       if (section && section.type === 'userText') {
         this.angularFireAuth.user
-          .pipe(switchMap(user => {
-            return user ? this.angularFirestore.collection('user_account_tap').doc(user.uid).valueChanges() : of(null);
-          }))
-          .subscribe(user => {
-            this.user = user;
-            section.body = this.fillUserData(section.body);
-          });
+        .pipe(switchMap(user => {
+          return user ? this.angularFirestore.collection('user_account_tap').doc(user.uid).valueChanges() : of(null);
+        }))
+        .subscribe(user => {
+          this.user = user;
+          section.body = this.fillUserData(section.body);
+        });
       }
     }
     return sections