+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add screen to update ID and trigger point synchronization processes (https://abi-consumerdigitalproducts.atlassian.net/browse/MAL-174)
 - Changed Drupal API endpoint and adapted to new response structure
 - Added a static Sitemap to the Home page (https://github.com/ab-inbev-global-martech/TapIt-Web/)
+- Logged/not logged version of the page is rendered by SSR
 ### Changed
 - CTA for internal sections won't change URL in order to avoid redirection on user login
 - Fixed the description on banner sections that wasn't displaying the copy text